Add facts for classification of operating systems
Add two new facts: ostype and osfamily. ostype is meant to be the same on
closely related operating systems which are likely to have similar
configuration. For example, most package names and configuration files are the
same on RedHat, Fedora, and CentOS, so it is useful to give all of these
ostype = "RedHat". SuSE, OpenSuSE, SLED, and SLES are similarly related.

osfamily is meant to be the same on distantly related systems which share
common concepts. For example, in a large, heterogeneous organization, one
might want to set up /etc/resolv.conf identically on every kind of Unix system
without having to enumerate every kernel type that is or may be deployed in
the puppet manifest.